Refactor: TableView, CellView 네이밍 구분
