+++ ATSOPT36_Assignment/ATSOPT36_Assignment/Presentation/Main/TodayTving/TodayTvingCell.swift
@@ -1,5 +1,5 @@
 //
-//  TodayTvingCell.swift
+//  TodayTvingCollectionCell.swift
 //  ATSOPT36_Assignment
 //
 //  Created by 권석기 on 5/1/25.
@@ -9,102 +9,48 @@ import UIKit
 
 import SnapKit
 
-final class TodayTvingCell: UITableViewCell {
-  
-    enum Metric {
-        static let itemSize: CGSize = CGSize(width: 150, height: 146)
-        static let itemMinimumSpacing: CGFloat = 11.0
-        static let itemMinimumInterSpacing: CGFloat = 11.0
-        static let sectionInset = UIEdgeInsets(top: 0, left: 10, bottom: 0, right: 0)
-    }
-    
-    // MARK: - Properties
-    private var items: [ContentModel] = []
+final class TodayTvingCell: UICollectionViewCell {
     
     static let identifier = "TodayTvingCell"
     
-    private let collectionViewHeader = TvingCollectionHeaderView().then {
-        $0.configure(title: "오늘의 티빙 TOP 20")
-        $0.backgroundColor = .black
+    private let imageView = UIImageView().then {
+        $0.layer.cornerRadius = 3
+        $0.clipsToBounds = true
+        $0.image = UIImage(resource: .movie1)
     }
     
-    private lazy var collectionViewLayout = UICollectionViewFlowLayout().then {
-        $0.scrollDirection = .horizontal
-        $0.minimumLineSpacing = Metric.itemMinimumSpacing
-        $0.minimumInteritemSpacing = Metric.itemMinimumInterSpacing
-        $0.itemSize = Metric.itemSize
-        $0.sectionInset = Metric.sectionInset
+    private let rankingLabel = UIImageView().then {
+        $0.image = UIImage(resource: .number1)
     }
     
-    private lazy var collectionView = UICollectionView(frame: .zero, collectionViewLayout: collectionViewLayout).then {
-        $0.showsHorizontalScrollIndicator = false
-        $0.register(TodayTvingCollectionCell.self, forCellWithReuseIdentifier: TodayTvingCollectionCell.identifier)
-    }
-    
-    // MARK: - Initializer
-    
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
+    override init(frame: CGRect) {
+        super.init(frame: frame)
         setUI()
-        setDelegate()
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-    // MARK: - UI Setting
-    
-    private func setUI() {
-        [collectionViewHeader, collectionView].forEach {
+    private func setUI() {      
+        [rankingLabel, imageView].forEach {
             contentView.addSubview($0)
         }
         
-        collectionViewHeader.snp.makeConstraints {
-            $0.leading.trailing.top.equalToSuperview()
-            $0.height.equalTo(50)
+        rankingLabel.snp.makeConstraints {
+            $0.leading.bottom.equalToSuperview()
+        }
+        imageView.snp.makeConstraints {
+            $0.leading.equalTo(rankingLabel.snp.trailing).offset(10)
+            $0.top.trailing.bottom.equalToSuperview()
         }
         
-        collectionView.snp.makeConstraints {
-            $0.leading.trailing.bottom.equalToSuperview()
-            $0.top.equalTo(collectionViewHeader.snp.bottom)
-        }
-    }
-    
-    private func setDelegate() {
-        collectionView.dataSource = self
-        collectionView.delegate = self
-    }
-    
-    func prepare(items: [ContentModel]) {
-        self.items = items
     }
 }
 
-// MARK: - UICollectionViewDataSource
-
-extension TodayTvingCell: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return items.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: TodayTvingCollectionCell.identifier, for: indexPath) as? TodayTvingCollectionCell else { return UICollectionViewCell() }
-        cell.dataBind(index: indexPath.row, items[indexPath.row])
-        return cell
+extension TodayTvingCell {
+    func dataBind(index: Int, _ item: ContentModel) {
+        imageView.image = item.thumbnail        
+        rankingLabel.image = UIImage(named: "number\(index + 1)")
     }
 }
-
-extension TodayTvingCell: UICollectionViewDelegateFlowLayout {
-    func scrollViewWillEndDragging(_ scrollView: UIScrollView, withVelocity velocity: CGPoint, targetContentOffset: UnsafeMutablePointer<CGPoint>) {
-        if !scrollView.isDragging {
-            let scrollOffset = scrollView.contentOffset.x
-            let rawIndex = scrollOffset / Metric.itemSize.width
-            let index: CGFloat
-            index = rawIndex - floor(rawIndex) > 0.7 ? CGFloat(Int(rawIndex + 1)) : CGFloat(Int(rawIndex))
-            let offset = (index * Metric.itemSize.width) + (index * Metric.itemMinimumSpacing)
-            targetContentOffset.pointee = CGPoint(x: offset, y: scrollView.contentInset.top)
-        }
-    }
-}
-
